Code cleanup: extract code of reservoir sample to separate class
